LayoutTests/fast/text/font-face-set-remove-safely-expected.txt

CONSOLE MESSAGE: Retrieving fonts...
CONSOLE MESSAGE: Clearing...
CONSOLE MESSAGE: Done.
